[level 2] Title: 전화번호 목록, Time: 76.50 ms, Memory: 35.6 MB -BaekjoonHub

+++ 프로그래머스/2/42577. 전화번호 목록/전화번호 목록.cpp
@@ -6,7 +6,7 @@ using namespace std;
 bool solution(vector<string> phone_book) {
     sort(phone_book.begin(), phone_book.end());
 
-    for (size_t i = 0; i < phone_book.size() - 1; ++i) {
+    for (int i = 0; i < phone_book.size() - 1; ++i) {
         if (phone_book[i] == phone_book[i + 1].substr(0, phone_book[i].length())) {
             return false; // 접두어 발견
         }
